First attempt at splitting up function.py
