+++ trieste/acquisition/multi_objective/__init__.py
@@ -11,8 +11,15 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-""" This folder contains multi-objective optimization utilities. """
+""" This folder contains multi-objective optimization functions and utilities. """
 from .dominance import non_dominated
+from .function import (
+    BatchMonteCarloExpectedHypervolumeImprovement,
+    ExpectedConstrainedHypervolumeImprovement,
+    ExpectedHypervolumeImprovement,
+    batch_ehvi,
+    expected_hv_improvement,
+)
 from .pareto import Pareto, get_reference_point
 from .partition import (
     DividedAndConquerNonDominated,
